pyproject.toml
[project]
name = "geodjango"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-stubs[compatible-mypy]>=5.2.8",
    "djangorestframework>=3.16.1",
    "djangorestframework-gis>=1.2.0",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs[compatible-mypy]>=3.16.6",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "psycopg[binary]>=3.3.2",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "ruff>=0.14.10",
    "wagtail>=7.3",
    "wagtailgeowidget>=9.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.5.0",
    "ruff>=0.14.4",
]

[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

fixable = ["ALL"]
unfixable = []

extend-select = ["E501"]

[tool.ruff.lint.isort]
combine-as-imports = true
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE="config.settings"

[tool.mypy]
python_version = "3.14"
warn_return_any = true
ignore_missing_imports = true
strict_optional = true
disallow_untyped_defs = true
check_untyped_defs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main"
]

[tool.django-stubs]
django_settings_module = "config.settings"